+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/DeltaReportGeneratorFacade.kt
@@ -5,22 +5,28 @@ import io.github.surpsg.deltacoverage.report.summary.CoverageCheckSummary
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.lang.invoke.MethodHandles
-import java.nio.file.Path
 
 abstract class DeltaReportGeneratorFacade {
 
     private val log: Logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass())
 
     fun generateReports(
-        summaryFileLocation: Path,
         config: DeltaCoverageConfig,
     ) {
         log.debug("[{}] Run Delta-Coverage with config: {}", config.view, config)
-        val coverageSummary: CoverageSummary = generate(ReportContext(config))
+        val coverageSummaries: Map<ReportBound, CoverageSummary> = generate(ReportContext(config))
 
-        CoverageCheckSummary.create(summaryFileLocation, coverageSummary)
-        CoverageViolationsPropagator().propagateAll(coverageSummary)
+        coverageSummaries.forEach { reportBound, coverageSummary ->
+            CoverageCheckSummary.create(
+                config.reportsConfig.summaries.getValue(reportBound),
+                coverageSummary,
+            )
+        }
+
+        CoverageViolationsPropagator().propagateAll(
+            coverageSummaries.getValue(ReportBound.DELTA_REPORT),
+        )
     }
 
-    internal abstract fun generate(reportContext: ReportContext): CoverageSummary
+    internal abstract fun generate(reportContext: ReportContext): Map<ReportBound, CoverageSummary>
 }